Basic address page with abstracted components
