+++ src/App.js
@@ -14,7 +14,7 @@ import SearchResults from './routes/SearchResults'
 import SingleName from './routes/SingleName'
 import Favourites from './routes/Favourites'
 import About from './routes/About'
-import Address from './routes/Address'
+import Address from './routes/AddressPage'
 import { TestPage } from './routes/Test'
 import Modal from './components/Modal/Modal'
 import Confirm from './components/SingleName/Confirm'
